test fixes/minor structure change + removed knex promises
