+++ README.md
@@ -17,14 +17,33 @@
 
 - Implement testing: Mocha/Chai/Supertest
 
-##### Run in Development:
+##### First things first
 
 - npm install
+
+##### Database
+
+- initdb db/
+
+- postgres -D db/ (or what they show you in Terminal, if it's different)
+
+- createdb DB_NAME
+
+- psql ss_dev (db meddling)
+
+
+##### Server
+
+- createdb ss_dev
 
 - npm start
 
 ##### Tests
 
-- npm test 
+Initial
+- dropdb ss_test
 
-- npm run [nyan, min, ]
+- createdb ss_test
+
+Consecutive
+- npm [test, run nyan, run spec]
